requirements.txt
boto3 ~= 1.35
sentry-sdk ~= 2.13
python-dotenv ~= 1.0
requests ~= 2.32
